Refactor content list into separate component
